Update getting-started-with-cpp-driver.txt
According to Andrew Morrow the nightly build is broken and has been for some time, it if worked, it was by mistake.  So I am sticking in a recommendation to use the full build.  That works and there are several outstanding forum questions about why the stand alone driver doesn't work/doesn't conform to docs.