Note availability of new C++ driver repository=
Add a note describing the availability of the new repository and link to the blog post explaining the change. The existing build instructions are adequate for the repository in its current state.

Reviewer: please note that I made this edit with the github editor tool, so I haven't verified rendering it.